+++ lib/main/storages/silent_mode_storage.dart
@@ -1,0 +1,22 @@
+import 'package:anchwatt/main/storages/prefs_storage.dart';
+
+class SilentModeStorage {
+  /* Static variables */
+
+  static const String _keyEnabled = 'silent_mode.enabled';
+
+  /* Variables */
+
+  final PrefsStorage _prefsStorage = PrefsStorage();
+
+  /* Methods */
+
+  Future<void> init() => _prefsStorage.init();
+
+  bool readEnabled() => _prefsStorage.readBool(key: _keyEnabled);
+
+  Future<void> writeEnabled(bool value) => _prefsStorage.writeBool(
+    key: _keyEnabled,
+    value: value,
+  );
+}