+++ lib/styles/borders.dart
@@ -7,5 +7,4 @@ const BorderRadius borderRadiusOptionsButton = BorderRadius.all(Radius.circular(
 const BorderRadius borderRadiusOptionsDialog = BorderRadius.all(Radius.circular(12));
 const BorderRadius borderRadiusOptionsModePill = BorderRadius.all(Radius.circular(6));
 const BorderRadius borderRadiusOptionsUpdateZone = BorderRadius.all(Radius.circular(8));
-const BorderRadius borderRadiusSoundModePill = BorderRadius.all(Radius.circular(6));
 const BorderRadius borderRadiusXpBar = BorderRadius.all(Radius.circular(4));